[DEV-3063] Refactoring
