+++ src/subdomains/supporting/bank/bank-account/dto/iban.dto.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsNotEmpty } from 'class-validator';
 import { Util } from 'src/shared/utils/util';
-import { UpdateUserBankDataDto } from 'src/subdomains/generic/user/models/bank-data/dto/update-bank-data.dto';
+import { UpdateBankAccountDto } from 'src/subdomains/generic/user/models/bank-data/dto/update-bank-data.dto';
 import { IbanType, IsDfxIban } from '../is-dfx-iban.validator';
 
 export class IbanDto {
@@ -10,7 +10,7 @@ export class IbanDto {
   iban: string;
 }
 
-export class CreateIbanDto extends UpdateUserBankDataDto {
+export class CreateIbanDto extends UpdateBankAccountDto {
   @ApiProperty()
   @IsNotEmpty()
   @IsDfxIban(IbanType.BOTH)
