--- src/subdomains/supporting/bank/bank-account/dto/update-bank-account.dto.ts
@@ -1,23 +1,0 @@
-import { ApiPropertyOptional } from '@nestjs/swagger';
-import { Type } from 'class-transformer';
-import { IsBoolean, IsOptional, IsString, ValidateNested } from 'class-validator';
-import { EntityDto } from 'src/shared/dto/entity.dto';
-import { Fiat } from 'src/shared/models/fiat/fiat.entity';
-
-export class UpdateBankAccountDto {
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsString()
-  label: string;
-
-  @ApiPropertyOptional({ type: EntityDto })
-  @IsOptional()
-  @ValidateNested()
-  @Type(() => EntityDto)
-  preferredCurrency: Fiat;
-
-  @ApiPropertyOptional()
-  @IsOptional()
-  @IsBoolean()
-  active = true;
-}
